PLantList component changes, useForm change and CheckoutForm changes, no testing
